# Use std::ranges::subrange instead of boost::iterator_range