--- a/include/customizer/cell_customizer.hpp
+++ b/include/customizer/cell_customizer.hpp
@@ -100,9 +100,9 @@ class CellCustomizer
                 distances.front() =
                     inserted ? heap.GetData(destination).distance : INVALID_EDGE_DISTANCE;
 
-                weights.advance_begin(1);
-                durations.advance_begin(1);
-                distances.advance_begin(1);
+                weights.advance(1);
+                durations.advance(1);
+                distances.advance(1);
             }
             BOOST_ASSERT(weights.empty());
             BOOST_ASSERT(durations.empty());