--- a/include/engine/base64.hpp
+++ b/include/engine/base64.hpp
@@ -13,7 +13,6 @@
 #include <boost/archive/iterators/base64_from_binary.hpp>
 #include <boost/archive/iterators/binary_from_base64.hpp>
 #include <boost/archive/iterators/transform_width.hpp>
-#include <boost/range/algorithm/copy.hpp>
 
 namespace osrm
 {