--- a/include/extractor/class_data.hpp
+++ b/include/extractor/class_data.hpp
@@ -6,6 +6,7 @@
 #include <algorithm>
 #include <cctype>
 #include <cstdint>
+#include <string>
 
 namespace osrm::extractor
 {